requirements.txt
opencv-python
numpy
Pillow
imagehash
scipy
